+++ src/pyday/__init__.py
@@ -1,4 +1,5 @@
 from .ChangLanguage import *
 from .DataReader.DataReader import *
-from .Data2PDF.Data2PDF import *
+# from .Data2PDF.Data2PDF import *
+from .Data2PDF.Data2PDF2 import *
 from .DataVis.DataVis import *