+++ src/test.py
@@ -1,7 +1,3 @@
-from pyday import DataReader
-# from pyday import DataVis
 from pyday import Data2PDF
-from pyday import ChangLang
-
-d2pdf = Data2PDF("test.json")
-d2pdf.toFile("test.pdf")
+d2pdf = Data2PDF("test2.json")
+d2pdf.toFile("test2.pdf")